requirements.txt
asgiref==3.6.0
Django==4.2
Pillow==9.5.0
psycopg2==2.9.6
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
